Collapse all back to original place
