+++ server/src/test/java/org/eclipse/openvsx/admin/AdminAPITest.java
@@ -50,7 +50,6 @@ import org.eclipse.openvsx.UpstreamRegistryService;
 import org.eclipse.openvsx.UserService;
 import org.eclipse.openvsx.accesstoken.AccessTokenConfig;
 import org.eclipse.openvsx.accesstoken.AccessTokenService;
-import org.eclipse.openvsx.accesstoken.PersonalAccessTokens;
 import org.eclipse.openvsx.adapter.VSCodeIdService;
 import org.eclipse.openvsx.cache.CacheService;
 import org.eclipse.openvsx.cache.LatestExtensionVersionCacheKeyGenerator;
@@ -2216,19 +2215,12 @@ class AdminAPITest {
 
         @Bean
         AccessTokenService tokenService(
-                PersonalAccessTokens personalAccessTokens
-        ) {
-            return new AccessTokenService(personalAccessTokens);
-        }
-
-        @Bean
-        PersonalAccessTokens personalAccessTokens(
                 AccessTokenConfig config,
                 EntityManager entityManager,
                 RepositoryService repositories,
                 MailService mailService
         ) {
-            return new PersonalAccessTokens(config, entityManager, repositories, mailService);
+            return new AccessTokenService(config, entityManager, repositories, mailService);
         }
 
         @Bean
